==> 2_nn_feature_creation.ipynb <==
# Import the required libraries
import numpy as np
import pandas as pd
import hvplot.pandas
from pathlib import Path
from finta import TA
import plotly.graph_objects as go

# Import cleaned df
# Assure timestamp column is datetime
df = pd.read_csv("../algotrader2/resources/aapl_1min_df.csv")
display(df.head())

# First reset index
df.reset_index
# Set index to date-time
df['timestamp'] = pd.to_datetime(df['timestamp'], errors='coerce')
df = df.set_index("timestamp")
df.head()

# 1-minute indicator creation
# Percent Returns
df["pct_returns"] = df["close"].pct_change()
# MOVING AVERAGES

# Simple Moving Averages
df["SMA4"] = TA.SMA(df,4)
df["SMA100"] = TA.SMA(df,100)
df["SMA20"] = TA.SMA(df,20)
df["SMA50"] = TA.SMA(df,50)
df["SMA200"] = TA.SMA(df,200)

# Exponential Moving Averages
df["EMA3"] = TA.EMA(df,3)
df["EMA9"] = TA.EMA(df,9)
df["EMA20"] = TA.EMA(df,20)
df["EMA50"] = TA.EMA(df,50)
df["EMA100"] = TA.EMA(df,100)

# Double exponential moving average
df["DEMA3"] = TA.DEMA(df, 3)
df["DEMA7"] = TA.DEMA(df, 7)
df["DEMA13"] = TA.DEMA(df, 13)
df["DEMA15"] = TA.DEMA(df, 15)
df["DEMA17"] = TA.DEMA(df, 17)
df["DEMA21"] = TA.DEMA(df, 21)
df["DEMA50"] = TA.DEMA(df, 50)
df["DEMA77"] = TA.DEMA(df, 77)
df["DEMA80"] = TA.DEMA(df, 80)
df["DEMA99"] = TA.DEMA(df, 99)
df["DEMA150"] = TA.DEMA(df, 150)

# Triple exponential moving average
df["TEMA3"] = TA.TEMA(df, 3)
df["TEMA7"] = TA.TEMA(df, 7)
df["TEMA13"] = TA.TEMA(df, 13)
df["TEMA15"] = TA.TEMA(df, 15)
df["TEMA17"] = TA.TEMA(df, 17)
df["TEMA21"] = TA.TEMA(df, 21)
df["TEMA50"] = TA.TEMA(df, 50)
df["TEMA77"] = TA.TEMA(df, 77)
df["TEMA80"] = TA.TEMA(df, 80)
df["TEMA99"] = TA.TEMA(df, 99)
df["TEMA150"] = TA.TEMA(df, 150)

# Triangular exponential moving average
df["TRIMA"] = TA.TRIMA(df)
df["TRIMA3"] = TA.TRIMA(df, 3)
df["TRIMA6"] = TA.TRIMA(df, 6)
df["TRIMA9"] = TA.TRIMA(df, 9)
df["TRIMA15"] = TA.TRIMA(df, 15)
df["TRIMA21"] = TA.TRIMA(df, 21)
df["TRIMA27"] = TA.TRIMA(df, 27)
df["TRIMA66"] = TA.TRIMA(df, 66)
df["TRIMA81"] = TA.TRIMA(df, 81)
df["TRIMA99"] = TA.TRIMA(df, 99)
df["TRIMA150"] = TA.TRIMA(df, 150)

# * Triple Exponential Moving Average Oscillator 'TRIX'
df["TRIX"] = TA.TRIX(df)

# * Volume Adjusted Moving Average 'VAMA'
df["VAMA"] = TA.VAMA(df)

# Kaufman Efficiency Indicator
df["ER"] = TA.ER(df)

# Zero Lag Exponential Moving Average
df["ZLEMA"] = TA.ZLEMA(df)

# Weighted Moving Average
df["WMA"] = TA.WMA(df)

# Hull Moving Average
df["HMA"] = TA.HMA(df)

# Volume Weighted Average Price
df["VWAP"] = TA.VWAP(df)

# Smoothed Moving Average
df["SMMA"] = TA.SMMA(df)

# Fractal Adaptive Moving Average
df["FRAMA"] = TA.FRAMA(df)

# Market Momentum
df["MOM"] = TA.MOM(df)

# Rate-of-Change
df["ROC"] = TA.ROC(df)

# Relative Strength Index
df["RSI"] = TA.RSI(df)

# Inverse Fisher Transform RSI
df["IFT_RSI"] = TA.IFT_RSI(df)

# True Range
df["TR"] = TA.TR(df)

# Average True Range
df["ATR"] = TA.ATR(df)

# Stop-and-Reverse
df["SAR"] = TA.SAR(df)

# Bollinger Bands Width
df["BBWIDTH"] = TA.BBWIDTH(df)

# Percent B
df["PERCENT_B"] = TA.PERCENT_B(df)

# Stochastic Oscillator %K
df["STOCH"] = TA.STOCH(df)

# Stochastic Oscillator %D
df["STOCHD"] = TA.STOCHD(df)

# Stochastic RSI
df["STOCHRSI"] = TA.STOCHRSI(df)

# Williams %R
df["WILLIAMS"] = TA.WILLIAMS(df)

# Ultimate Oscillator
df["UO"] = TA.UO(df)

# Awesome Oscillator
df["AO"] = TA.AO(df)

# Mass Index
df["MI"] = TA.MI(df)

# Typical Price
df["TP"] = TA.TP(df)

# Accumulation-Distribution Line
df["ADL"] = TA.ADL(df)

# Chaikin Oscillator
df["CHAIKIN"] = TA.CHAIKIN(df)

# Money Flow Index
df["MFI"] = TA.MFI(df)

# On Balance Volume
df["OBV"] = TA.OBV(df)

# Weighter OBV
df["WOBV"] = TA.WOBV(df)

# Volume Zone Oscillator
df["VZO"] = TA.VZO(df)

# Price Zone Oscillator
df["PZO"] = TA.PZO(df)

# Elder's Force Index
df["EFI"] = TA.EFI(df)

# Cumulative Force Index
df["CFI"] = TA.CFI(df)

# Ease of Movement
df["EMV"] = TA.EMV(df)

# Commodity Channel Index
df["CCI"] = TA.CCI(df)

# Coppock Curve
df["COPP"] = TA.COPP(df)

# Chande Momentum Oscillator
df["CMO"] = TA.CMO(df)

# Fisher Transform
df["FISH"] = TA.FISH(df)

# Squeeze Momentum Indicator
df["SQZMI"] = TA.SQZMI(df)

# Volume Price Trend
df["VPT"] = TA.VPT(df)

# Finite Volume Element
df["FVE"] = TA.FVE(df)

# Volume Flow Indicator
df["VFI"] = TA.VFI(df)

# Moving Standard Deviation
df["MSD"] = TA.MSD(df)

# Schaff Trend Cycle
df["STC"] = TA.STC(df)

# Drop NaN
df=df.dropna()

df.head()

(for future updates)
SINGLE OUTPUTS, RUNNING WITH DEFAULT TEMPLATE
Need to put some thought into each of the different indicators. If we can add values like "TA.ER(df, 10, 2, Adjust = false)", will that improve our algo? Probably so. What about doing something similar to what we did with the moving averages?

# updates to come...
## INDICATORS THAT GIVE TWO OUTPUT columns
# We will go back and add these when there is more time

# Kaufman's Adaptive Moving Average
#df["KAMA"] = TA.KAMA(df,10)

# Elastic Volume Moving Average
#df["EVWMA"] = TA.EVWMA(df)

# Moving Average Convergence Divergence
#df["MACD"] = TA.MACD(df)

# Percentage Price Oscillator
#df["PPO"] = TA.PPO(df)

# Volume-Weighted MACD
# df["VW_MACD"] = TA.VW_MACD(df)

# Elastic-Volume weighted MACD
# df["EV_MACD"] = TA.EV_MACD(df)

# Momentum Breakout Bands
#df["MOBO"] = TA.MOBO(df)

# Bollinger Bands
#df["BBANDS"] = TA.BBANDS(df)

# Keltner Channels
#df["KC"] = TA.KC(df)

# Donchian Channel
# df["DO"] = TA.DO(df)

# Directional Movement Indicator
# df["DMI"] = TA.DMI(df)


# Pivot Points
# df["PIVOT"] = TA.PIVOT(df)


# Fibonacci Pivot Points
# df["PIVOT_FIB"] = TA.PIVOT_FIB(df)

# Vortex Indicator
# df["VORTEX"] = TA.VORTEX(df)

# Know Sure Thing
#df["KST"] = TA.KST(df)

# True Strength Index
#df["TSI"] = TA.TSI(df)

# Bull power and Bear Power
# df["EBBP"] = TA.EBBP(df)

# Buy and Sell Pressure
# df["BASP"] = TA.BASP(df)

# Normalized BASP
# df["BASPN"] = TA.BASPN(df)

# Chandelier Exit
# df["CHANDELIER"] = TA.CHANDELIER(df)


# # Wave Trend Oscillator
# df["WTO"] = TA.WTO(df)

# Ichimoku Cloud
# df["ICHIMOKU"] = TA.ICHIMOKU(df)

# Adaptive Price Zone
# df["APZ"] = TA.APZ(df)

# OTHER ISSUES

# Twiggs Money Index
#df["TMF"] = TA.TMF(df)

# Reset the index for the following code
df = df.reset_index()

# Call in isPivot function
# "function that detects if a candle is a pivot/fractal point
# args: candle index, window before and after candle to test if pivot
# returns: 1 if pivot high, -1 if pivot low, 0 if both and 0 default"
from isPivotFunction import isPivot 
isPivot

# Run isPivot function and store high and low points in new column with a window of +-5 candles 
window=3
df['isPivot'] = df.apply(lambda x: isPivot(df, x.name,window), axis=1)

# Create pointpos function to help visualize isPivot
def pointpos(x):
    if x['isPivot']==-1:
        return x['low']-1e-3
    elif x['isPivot']==1:
        return x['high']+1e-3
    else:
        return np.nan
df['pointpos'] = df.apply(lambda row: pointpos(row), axis=1)

# Plot candles and pivot points
dfpl = df
fig = go.Figure(data=[go.Candlestick(x=dfpl['timestamp'],
                open=dfpl['open'],
                high=dfpl['high'],
                low=dfpl['low'],
                close=dfpl['close'])])

fig.add_scatter(x=dfpl['timestamp'], y=dfpl['pointpos'], mode="markers",
                marker=dict(size=5, color="MediumPurple"),
                name="pivot")
fig.update_layout(
    title_text="AAPL Chart with Pivot Points",
    xaxis_title="Date",
    yaxis_title="Price",
    xaxis_rangeslider_visible=False
)
fig.show()

# Drop pointpos columns
df.drop(columns='pointpos', inplace=True)
# Rename isPivot column to signal
df.rename(columns={'isPivot': 'signal'}, inplace=True)
df.set_index('timestamp', inplace=True)
# Review dataframe
df.head()

# We actually want the signal column to just be one of two things, 0 for undervalued, 1 for overvalued.
df['new_signal'] = 1
# Initialize counter
x = 0 
for x in range(len(df)):
    if df['signal'][x] == 1:
        df['new_signal'][x] = 1
    elif df['signal'][x] == -1:
        df['new_signal'][x] = 0
    else:
        df['new_signal'][x] = df['new_signal'][x-1]

display(df.head())
# Drop old signal
df.drop(columns='signal', inplace=True)
# Now save this dataframe to a new .csv
df.to_csv('../algotrader2/Resources/aapl_1min_indc_df.csv', index=True)

### Using 1-minute candles with the pivot points looks very overfitted.


Let's do the same as above with 3 minute, and then 15 minute candles. Then we will implement our neural network.

# 3-minute indicator creation

# Import cleaned df
df = pd.read_csv("../algotrader2/resources/aapl_3min_df.csv")

# Assure timestamp column is datetime
# First reset index
df.reset_index
df['timestamp'] = pd.to_datetime(df['timestamp'], errors='coerce')

df = df.set_index("timestamp")

# Create indicators
df["pct_returns"] = df["close"].pct_change()

# MOVING AVERAGES

# Simple Moving Averages
df["SMA5"] = TA.SMA(df,5)
df["SMA10"] = TA.SMA(df,10)
df["SMA20"] = TA.SMA(df,20)
df["SMA50"] = TA.SMA(df,50)
df["SMA200"] = TA.SMA(df,200)

# Exponential Moving Averages
df["EMA3"] = TA.EMA(df,3)
df["EMA9"] = TA.EMA(df,9)
df["EMA20"] = TA.EMA(df,20)
df["EMA50"] = TA.EMA(df,50)
df["EMA100"] = TA.EMA(df,100)

# Double exponential moving average
df["DEMA3"] = TA.DEMA(df, 3)
df["DEMA7"] = TA.DEMA(df, 7)
df["DEMA13"] = TA.DEMA(df, 13)
df["DEMA15"] = TA.DEMA(df, 15)
df["DEMA17"] = TA.DEMA(df, 17)
df["DEMA21"] = TA.DEMA(df, 21)
df["DEMA50"] = TA.DEMA(df, 50)
df["DEMA77"] = TA.DEMA(df, 77)
df["DEMA80"] = TA.DEMA(df, 80)
df["DEMA99"] = TA.DEMA(df, 99)
df["DEMA150"] = TA.DEMA(df, 150)

# Triple exponential moving average
df["TEMA3"] = TA.TEMA(df, 3)
df["TEMA7"] = TA.TEMA(df, 7)
df["TEMA13"] = TA.TEMA(df, 13)
df["TEMA15"] = TA.TEMA(df, 15)
df["TEMA17"] = TA.TEMA(df, 17)
df["TEMA21"] = TA.TEMA(df, 21)
df["TEMA50"] = TA.TEMA(df, 50)
df["TEMA77"] = TA.TEMA(df, 77)
df["TEMA80"] = TA.TEMA(df, 80)
df["TEMA99"] = TA.TEMA(df, 99)
df["TEMA150"] = TA.TEMA(df, 150)

# Triangular exponential moving average
df["TRIMA"] = TA.TRIMA(df)
df["TRIMA3"] = TA.TRIMA(df, 3)
df["TRIMA6"] = TA.TRIMA(df, 6)
df["TRIMA9"] = TA.TRIMA(df, 9)
df["TRIMA15"] = TA.TRIMA(df, 15)
df["TRIMA21"] = TA.TRIMA(df, 21)
df["TRIMA27"] = TA.TRIMA(df, 27)
df["TRIMA66"] = TA.TRIMA(df, 66)
df["TRIMA81"] = TA.TRIMA(df, 81)
df["TRIMA99"] = TA.TRIMA(df, 99)
df["TRIMA150"] = TA.TRIMA(df, 150)

# * Triple Exponential Moving Average Oscillator 'TRIX'
df["TRIX"] = TA.TRIX(df)

# * Volume Adjusted Moving Average 'VAMA'
df["VAMA"] = TA.VAMA(df)

# Kaufman Efficiency Indicator
df["ER"] = TA.ER(df)

# Zero Lag Exponential Moving Average
df["ZLEMA"] = TA.ZLEMA(df)

# Weighted Moving Average
df["WMA"] = TA.WMA(df)

# Hull Moving Average
df["HMA"] = TA.HMA(df)

# Volume Weighted Average Price
df["VWAP"] = TA.VWAP(df)

# Smoothed Moving Average
df["SMMA"] = TA.SMMA(df)

# Fractal Adaptive Moving Average
df["FRAMA"] = TA.FRAMA(df)

# Market Momentum
df["MOM"] = TA.MOM(df)

# Rate-of-Change
df["ROC"] = TA.ROC(df)

# Relative Strength Index
df["RSI"] = TA.RSI(df)

# Inverse Fisher Transform RSI
df["IFT_RSI"] = TA.IFT_RSI(df)

# True Range
df["TR"] = TA.TR(df)

# Average True Range
df["ATR"] = TA.ATR(df)

# Stop-and-Reverse
df["SAR"] = TA.SAR(df)

# Bollinger Bands Width
df["BBWIDTH"] = TA.BBWIDTH(df)

# Percent B
df["PERCENT_B"] = TA.PERCENT_B(df)

# Stochastic Oscillator %K
df["STOCH"] = TA.STOCH(df)

# Stochastic Oscillator %D
df["STOCHD"] = TA.STOCHD(df)

# Stochastic RSI
df["STOCHRSI"] = TA.STOCHRSI(df)

# Williams %R
df["WILLIAMS"] = TA.WILLIAMS(df)

# Ultimate Oscillator
df["UO"] = TA.UO(df)

# Awesome Oscillator
df["AO"] = TA.AO(df)

# Mass Index
df["MI"] = TA.MI(df)

# Typical Price
df["TP"] = TA.TP(df)

# Accumulation-Distribution Line
df["ADL"] = TA.ADL(df)

# Chaikin Oscillator
df["CHAIKIN"] = TA.CHAIKIN(df)

# Money Flow Index
df["MFI"] = TA.MFI(df)

# On Balance Volume
df["OBV"] = TA.OBV(df)

# Weighter OBV
df["WOBV"] = TA.WOBV(df)

# Volume Zone Oscillator
df["VZO"] = TA.VZO(df)

# Price Zone Oscillator
df["PZO"] = TA.PZO(df)

# Elder's Force Index
df["EFI"] = TA.EFI(df)

# Cumulative Force Index
df["CFI"] = TA.CFI(df)

# Ease of Movement
df["EMV"] = TA.EMV(df)

# Commodity Channel Index
df["CCI"] = TA.CCI(df)

# Coppock Curve
df["COPP"] = TA.COPP(df)

# Chande Momentum Oscillator
df["CMO"] = TA.CMO(df)

# Fisher Transform
df["FISH"] = TA.FISH(df)

# Squeeze Momentum Indicator
df["SQZMI"] = TA.SQZMI(df)

# Volume Price Trend
df["VPT"] = TA.VPT(df)

# Finite Volume Element
df["FVE"] = TA.FVE(df)

# Volume Flow Indicator
df["VFI"] = TA.VFI(df)

# Moving Standard Deviation
df["MSD"] = TA.MSD(df)

# Schaff Trend Cycle
df["STC"] = TA.STC(df)

# Drop NaN
df=df.dropna()

# Reset the index for the following code
df.reset_index(inplace=True)

from isPivotFunction import isPivot 
# Run isPivot function and store high and low points in new column with a window of +-5 candles 
window=3
df['isPivot'] = df.apply(lambda x: isPivot(df, x.name,window), axis=1)

df.head()

# Create pointpos function to help visualize isPivot
def pointpos(x):
    if x['isPivot']==-1:
        return x['low']-1e-3
    elif x['isPivot']==1:
        return x['high']+1e-3
    else:
        return np.nan
df['pointpos'] = df.apply(lambda row: pointpos(row), axis=1)

# Plot candles and pivot points
dfpl = df
fig = go.Figure(data=[go.Candlestick(x=dfpl['timestamp'],
                open=dfpl['open'],
                high=dfpl['high'],
                low=dfpl['low'],
                close=dfpl['close'])])

fig.add_scatter(x=dfpl['timestamp'], y=dfpl['pointpos'], mode="markers",
                marker=dict(size=5, color="MediumPurple"),
                name="pivot")
fig.update_layout(
    title_text="3min AAPL Chart with Pivot Points",
    xaxis_title="Date",
    yaxis_title="Price",
    xaxis_rangeslider_visible=False
)
fig.show()

# Drop pointpos columns
df.drop(columns='pointpos', inplace=True)
# Rename isPivot column to signal
df.rename(columns={'isPivot': 'signal'}, inplace=True)
df.set_index('timestamp', inplace=True)

# We actually want to signal column to just be one of two things, 0 for undervalued, 1 for overvalued.
df['new_signal'] = 1
# Initialize counter
x = 0 
for x in range(len(df)):
    if df['signal'][x] == 1:
        df['new_signal'][x] = 1
    elif df['signal'][x] == -1:
        df['new_signal'][x] = 0
    else:
        df['new_signal'][x] = df['new_signal'][x-1]
# Drop old signal
df.drop(columns='signal', inplace=True)
# Now save this dataframe to a new .csv
df.to_csv('../algotrader2/Resources/aapl_3min_indc_df.csv', index=True)

# 15-minute indicator creation
df = pd.read_csv("../algotrader2/resources/aapl_15min_df.csv")
# Assure timestamp column is datetime
# First reset index
df.reset_index
df['timestamp'] = pd.to_datetime(df['timestamp'], errors='coerce')
df = df.set_index("timestamp")

# Create indicators
df["pct_returns"] = df["close"].pct_change()
# MOVING AVERAGES

# Simple Moving Averages
df["SMA5"] = TA.SMA(df,5)
df["SMA10"] = TA.SMA(df,10)
df["SMA20"] = TA.SMA(df,20)
df["SMA50"] = TA.SMA(df,50)
df["SMA200"] = TA.SMA(df,200)

# Exponential Moving Averages
df["EMA3"] = TA.EMA(df,3)
df["EMA9"] = TA.EMA(df,9)
df["EMA20"] = TA.EMA(df,20)
df["EMA50"] = TA.EMA(df,50)
df["EMA100"] = TA.EMA(df,100)

# Double exponential moving average
df["DEMA3"] = TA.DEMA(df, 3)
df["DEMA7"] = TA.DEMA(df, 7)
df["DEMA13"] = TA.DEMA(df, 13)
df["DEMA15"] = TA.DEMA(df, 15)
df["DEMA17"] = TA.DEMA(df, 17)
df["DEMA21"] = TA.DEMA(df, 21)
df["DEMA50"] = TA.DEMA(df, 50)
df["DEMA77"] = TA.DEMA(df, 77)
df["DEMA80"] = TA.DEMA(df, 80)
df["DEMA99"] = TA.DEMA(df, 99)
df["DEMA150"] = TA.DEMA(df, 150)

# Triple exponential moving average
df["TEMA3"] = TA.TEMA(df, 3)
df["TEMA7"] = TA.TEMA(df, 7)
df["TEMA13"] = TA.TEMA(df, 13)
df["TEMA15"] = TA.TEMA(df, 15)
df["TEMA17"] = TA.TEMA(df, 17)
df["TEMA21"] = TA.TEMA(df, 21)
df["TEMA50"] = TA.TEMA(df, 50)
df["TEMA77"] = TA.TEMA(df, 77)
df["TEMA80"] = TA.TEMA(df, 80)
df["TEMA99"] = TA.TEMA(df, 99)
df["TEMA150"] = TA.TEMA(df, 150)

# Triangular exponential moving average
df["TRIMA"] = TA.TRIMA(df)
df["TRIMA3"] = TA.TRIMA(df, 3)
df["TRIMA6"] = TA.TRIMA(df, 6)
df["TRIMA9"] = TA.TRIMA(df, 9)
df["TRIMA15"] = TA.TRIMA(df, 15)
df["TRIMA21"] = TA.TRIMA(df, 21)
df["TRIMA27"] = TA.TRIMA(df, 27)
df["TRIMA66"] = TA.TRIMA(df, 66)
df["TRIMA81"] = TA.TRIMA(df, 81)
df["TRIMA99"] = TA.TRIMA(df, 99)
df["TRIMA150"] = TA.TRIMA(df, 150)

# * Triple Exponential Moving Average Oscillator 'TRIX'
df["TRIX"] = TA.TRIX(df)

# * Volume Adjusted Moving Average 'VAMA'
df["VAMA"] = TA.VAMA(df)

# Kaufman Efficiency Indicator
df["ER"] = TA.ER(df)

# Zero Lag Exponential Moving Average
df["ZLEMA"] = TA.ZLEMA(df)

# Weighted Moving Average
df["WMA"] = TA.WMA(df)

# Hull Moving Average
df["HMA"] = TA.HMA(df)

# Volume Weighted Average Price
df["VWAP"] = TA.VWAP(df)

# Smoothed Moving Average
df["SMMA"] = TA.SMMA(df)

# Fractal Adaptive Moving Average
df["FRAMA"] = TA.FRAMA(df)

# Market Momentum
df["MOM"] = TA.MOM(df)

# Rate-of-Change
df["ROC"] = TA.ROC(df)

# Relative Strength Index
df["RSI"] = TA.RSI(df)

# Inverse Fisher Transform RSI
df["IFT_RSI"] = TA.IFT_RSI(df)

# True Range
df["TR"] = TA.TR(df)

# Average True Range
df["ATR"] = TA.ATR(df)

# Stop-and-Reverse
df["SAR"] = TA.SAR(df)

# Bollinger Bands Width
df["BBWIDTH"] = TA.BBWIDTH(df)

# Percent B
df["PERCENT_B"] = TA.PERCENT_B(df)

# Stochastic Oscillator %K
df["STOCH"] = TA.STOCH(df)

# Stochastic Oscillator %D
df["STOCHD"] = TA.STOCHD(df)

# Stochastic RSI
df["STOCHRSI"] = TA.STOCHRSI(df)

# Williams %R
df["WILLIAMS"] = TA.WILLIAMS(df)

# Ultimate Oscillator
df["UO"] = TA.UO(df)

# Awesome Oscillator
df["AO"] = TA.AO(df)

# Mass Index
df["MI"] = TA.MI(df)

# Typical Price
df["TP"] = TA.TP(df)

# Accumulation-Distribution Line
df["ADL"] = TA.ADL(df)

# Chaikin Oscillator
df["CHAIKIN"] = TA.CHAIKIN(df)

# Money Flow Index
df["MFI"] = TA.MFI(df)

# On Balance Volume
df["OBV"] = TA.OBV(df)

# Weighter OBV
df["WOBV"] = TA.WOBV(df)

# Volume Zone Oscillator
df["VZO"] = TA.VZO(df)

# Price Zone Oscillator
df["PZO"] = TA.PZO(df)

# Elder's Force Index
df["EFI"] = TA.EFI(df)

# Cumulative Force Index
df["CFI"] = TA.CFI(df)

# Ease of Movement
df["EMV"] = TA.EMV(df)

# Commodity Channel Index
df["CCI"] = TA.CCI(df)

# Coppock Curve
df["COPP"] = TA.COPP(df)

# Chande Momentum Oscillator
df["CMO"] = TA.CMO(df)

# Fisher Transform
df["FISH"] = TA.FISH(df)

# Squeeze Momentum Indicator
df["SQZMI"] = TA.SQZMI(df)

# Volume Price Trend
df["VPT"] = TA.VPT(df)

# Finite Volume Element
df["FVE"] = TA.FVE(df)

# Volume Flow Indicator
df["VFI"] = TA.VFI(df)

# Moving Standard Deviation
df["MSD"] = TA.MSD(df)

# Schaff Trend Cycle
df["STC"] = TA.STC(df)

# Drop NaN
df=df.dropna()

df

# Reset the index for the following code
df.reset_index(inplace=True)
from isPivotFunction import isPivot 
# Run isPivot function and store high and low points in new column with a window of +-5 candles 
window=3
df['isPivot'] = df.apply(lambda x: isPivot(df, x.name,window), axis=1)
# Create pointpos function to help visualize isPivot
def pointpos(x):
    if x['isPivot']==-1:
        return x['low']-1e-3
    elif x['isPivot']==1:
        return x['high']+1e-3
    else:
        return np.nan
df['pointpos'] = df.apply(lambda row: pointpos(row), axis=1)

# Plot candles and pivot points
dfpl = df
#dfpl = df[df['Date'].dt.year == 2020]
#dfpl = df[(df['Date'].dt.year >= 2020) & (df['Date'].dt.year <= 2023)]
fig = go.Figure(data=[go.Candlestick(x=dfpl['timestamp'],
                open=dfpl['open'],
                high=dfpl['high'],
                low=dfpl['low'],
                close=dfpl['close'])])

fig.add_scatter(x=dfpl['timestamp'], y=dfpl['pointpos'], mode="markers",
                marker=dict(size=5, color="MediumPurple"),
                name="pivot")
fig.update_layout(
    title_text="15min AAPL Chart with Pivot Points",
    xaxis_title="Date",
    yaxis_title="Price",
    xaxis_rangeslider_visible=False
)
fig.show()

# Drop pointpos columns
df.drop(columns='pointpos', inplace=True)
# Rename isPivot column to signal
df.rename(columns={'isPivot': 'signal'}, inplace=True)
# Review dataframe
display(df.head())
df.set_index('timestamp', inplace=True)

# We actually want to signal column to just be one of two things, 0 for undervalued, 1 for overvalued.
df['new_signal'] = 1
# Initialize counter
x = 0 
for x in range(len(df)):
    if df['signal'][x] == 1:
        df['new_signal'][x] = 1
    elif df['signal'][x] == -1:
        df['new_signal'][x] = 0
    else:
        df['new_signal'][x] = df['new_signal'][x-1]
# Drop old signal
df.drop(columns='signal', inplace=True)
# Now save this dataframe to a new .csv
df.to_csv('../algotrader2/Resources/aapl_15min_indc_df.csv', index=True)
