==> 7_nn_trade_api.ipynb <==
#### For future reference when deploying our algorithm.

# Set signal variable
signal = 1


# Create buy signal, num shares and ticker
if signal == 1:
    orderSide = "buy"
else:
    orderSide = "sell"


# Set the ticket symbol and the number of shares to buy
ticker = "AAPL"
number_of_shares = 1


# # Make API call
# prices = alpaca.get_bars(ticker, "1Min").df

# # Reorganize the DataFrame
# prices = pd.concat([prices], axis=1, keys=["AAPL"])

# # Get final closing price
# limit_amount = prices["AAPL"]["close"][-1]



# # Submit order
# alpaca.submit_order(
#     symbol="AAPL", 
#     qty=number_of_shares, 
#     side=orderSide, 
#     time_in_force="gtc", 
#     type="limit", 
#     limit_price=limit_amount
# )

def execute_trades(api, symbol, entry_model, exit_model, features):
    entry_signals = entry_model.predict(features)
    exit_signals = exit_model.predict(features)

    for i in range(1, len(entry_signals)):
        current_entry_signal = entry_signals[i]
        previous_entry_signal = entry_signals[i - 1]

        current_exit_signal = exit_signals[i]
        previous_exit_signal = exit_signals[i - 1]

        if current_entry_signal == 1 and previous_entry_signal == 0:
            api.submit_order(
                symbol=symbol,
                qty=1,
                side='buy',
                type='market',
                time_in_force='gtc'
            )
        elif current_entry_signal == -1 and previous_entry_signal == 0:
            api.submit_order(
                symbol=symbol,
                qty=1,
                side='sell',
                type='market',
                time_in_force='gtc'
            )

        if current_exit_signal == -1 and previous_exit_signal == 1:
            api.submit_order(
                symbol=symbol,
                qty=1,
                side='sell',
                type='market',
                time_in_force='gtc'
            )
